requirements.txt
requests 
pathlib 
pandas
python-dotenv
langchain
openai 
docarray
tiktoken 
cohere
pinecone-client